Extract vpn check options
